requirements.txt
asgiref==3.5.2
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
numpy==1.22.1
psycopg2==2.9.4
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
pyxlsb==1.0.9
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.8
whitenoise==6.2.0
pycryptodome
pyrebase4
ndg-httpsclient
